+++ integration-tests/auth-with-rpc/src/blitz-client.ts
@@ -1,0 +1,14 @@
+import {AuthClientPlugin} from "@blitzjs/auth"
+import {setupBlitzClient} from "@blitzjs/next"
+import {BlitzRpcPlugin} from "@blitzjs/rpc"
+import {BlitzCustomPlugin} from "./custom-plugin/plugin"
+
+export const {withBlitz} = setupBlitzClient({
+  plugins: [
+    AuthClientPlugin({
+      cookiePrefix: "web-cookie-prefix",
+    }),
+    BlitzRpcPlugin({}),
+    BlitzCustomPlugin({}),
+  ],
+})