+++ packages/blitz-rpc/src/index-browser.tsx
@@ -1,7 +1,6 @@
 import "./global"
 import {createClientPlugin} from "blitz"
 import {DefaultOptions, QueryClient} from "@tanstack/react-query"
-
 export * from "./data-client/index"
 
 interface BlitzRpcOptions {
@@ -36,9 +35,29 @@ export const BlitzRpcPlugin = createClientPlugin<BlitzRpcOptions, {queryClient: 
       })
     }
     const queryClient = initializeQueryClient()
+    function resetQueryClient() {
+      setTimeout(async () => {
+        // Do these in the next tick to prevent various bugs like https://github.com/blitz-js/blitz/issues/2207
+        const debug = (await import("debug")).default("blitz:rpc")
+        debug("Invalidating react-query cache...")
+        await queryClient.cancelQueries()
+        await queryClient.resetQueries()
+        queryClient.getMutationCache().clear()
+        // We have a 100ms delay here to prevent unnecessary stale queries from running
+        // This prevents the case where you logout on a page with
+        // Page.authenticate = {redirectTo: '/login'}
+        // Without this delay, queries that require authentication on the original page
+        // will still run (but fail because you are now logged out)
+        // Ref: https://github.com/blitz-js/blitz/issues/1935
+      }, 100)
+    }
     globalThis.queryClient = queryClient
     return {
-      events: {},
+      events: {
+        onSessionCreated: async () => {
+          resetQueryClient()
+        },
+      },
       middleware: {},
       exports: () => ({
         queryClient,