+++ recipes/next-ui/index.ts
@@ -1,8 +1,6 @@
 import {addImport, paths, Program, RecipeBuilder} from "blitz/installer"
 import type {NodePath} from "ast-types/lib/node-path"
 import j, {JSXIdentifier} from "jscodeshift"
-import fs from "fs"
-import path from "path"
 
 // Copied from https://github.com/blitz-js/legacy-framework/pull/805, let's add this to the blitz
 function wrapComponentWithChakraProvider(program: Program) {
@@ -113,7 +111,9 @@ export default RecipeBuilder()
     stepId: "updateLabeledTextField",
     stepName: "Update the `LabeledTextField` with Next UI's `Input` component",
     explanation: `The LabeledTextField component uses Next UI's input component`,
-    singleFileSearch: `${paths.appSrcDirectory()}/core/components/LabeledTextField.tsx`,
+    singleFileSearch: `${paths.appSrcDirectory({
+      isConfig: false,
+    })}/core/components/LabeledTextField.tsx`,
     transform(program) {
       // Add ComponentPropsWithoutRef import
       program.find(j.ImportDeclaration, {source: {value: "react"}}).forEach((path) => {